Ensure file descriptor safety for the restart coordination socket
When spawning the child process, we keep the RawFd of the restart
coordination socket. In fact the lifetime of that socket is distinct
from the thread that handles child spawning. This fact is not obvious
when the lifetime bounds are erased by taking the RawFd of the socket
listener.

To resolve this, dup the listener socket's file descriptor and hold onto
it in an OwnedFd for the lifetime of ChildSpawner's thread.