Implement handover via the systemd file descriptor store
Instead of forking the process and communicating with the child,
shellflip can communicate with systemd to send data in the 'old' process
and receive it in the 'new' process.

The restart can either be initiated through a Unix socket - in which case
the process and systemd unit file should agree on the exit code used to
restart the process - or by `systemctl restart`.

The process lifecycle is different when using systemd and its file
descriptor store. Systemd stops the process completely then spawns a new
one; old code does not stay running, unless you use the deprecated and
hated `KillMode=none` option.